Laravel 5.8 + Tailwind

+++ database/migrations/2018_01_15_052336_create_feeds_table.php
@@ -14,25 +14,25 @@ class CreateFeedsTable extends Migration
     public function up()
     {
         Schema::create('feeds', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('url', 180)->unique();
             $table->dateTime('last_check');
             $table->timestamps();
         });
         Schema::create('feed_subscriptions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedInteger('feed_id');
+            $table->unsignedBigInteger('feed_id');
             $table->foreign('feed_id')->references('id')->on('feeds');
             $table->unique(['user_id', 'feed_id']);
             $table->timestamps();
         });
         Schema::create('feed_posts', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title');
-            $table->unsignedInteger('feed_id');
+            $table->unsignedBigInteger('feed_id');
             $table->foreign('feed_id')->references('id')->on('feeds');
             $table->string('uuid', 180)->unique();
             $table->dateTime('created_at');
